Move objectAsFhir function to utils file
