+++ packages/shared-src/src/models/fhir/Resource.js
@@ -8,7 +8,7 @@ import {
   FHIR_SEARCH_TOKEN_TYPES,
   FHIR_DATETIME_PRECISION,
 } from '../../constants';
-import { objectAsFhir } from '../../utils/pgComposite';
+import { objectAsFhir } from './utils';
 import { formatFhirDate } from '../../utils/fhir';
 import { Model } from '../Model';
